+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [Unreleased]
+
+### Fixed
+
+* Fix implementation of `iterator#increment` in `csmmap` and `csmvector`.
+
+
 ## [4-r.1] - 2020-01-30
 
 ### Added
@@ -20,4 +27,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Reformat code using Prettier and ESLint.
 
 
+[Unreleased]: https://github.com/Live2D/CubismWebFramework/compare/4-r.1...HEAD
 [4-r.1]: https://github.com/Live2D/CubismWebFramework/compare/ce2585a919ac6e99f64dd468933772c6f1abbcc7...4-r.1